fix(prometheus): set wal_write_time and wal_sync_time as float64
Closes #2546

+++ pkg/management/postgres/probes.go
@@ -524,8 +524,8 @@ type PgStatWal struct {
 	WALBuffersFull int64
 	WalWrite       int64
 	WalSync        int64
-	WalWriteTime   int64
-	WalSyncTime    int64
+	WalWriteTime   float64
+	WalSyncTime    float64
 	StatsReset     string
 }
